Changes to input link
Previously "open" walls had a wme (<wall> ^to <to-id>) to indicate that a robot could travel from the current room to an adjacent room. This update removes that wme and replaces it with a (<wall> ^gateway <g>) structure similar to gateways that lead to doors. The difference is that where a room-door gateway has a wme (<g> ^room <room>) wme, the new room-room gateways have (<g> ^room nil).

Also added unique ids to walls.